requirements.txt
dash>=2.17
paho-mqtt>=1.6
pandas>=2.1
plotly>=5.20
PyYAML>=6.0
requests>=2.31
